requirements.txt
selenium
webdriver_manager
beautifulsoup4
django
Pillow
djangorestframework
django-allauth
djangorestframework-simplejwt
dj-rest-auth